requirements.txt
# Основные зависимости для Telegram бота
python-telegram-bot>=21.0,<22.0
pytz>=2023.3

# Для валидации данных и типизации
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Для работы с переменными окружения
python-dotenv>=1.0.0

# Дополнительные зависимости для разработки (опционально)
# black>=23.0.0  # Форматирование кода
# mypy>=1.0.0   # Проверка типов  
# pytest>=7.0.0 # Тестирование